cmr.rb: Noramlise issns for use with dataincubator.
periodicals.dataincubator.org expects URIs with a normalised issn tacked
on the end. Unixref sometimes throws us funny looking issns so we have
to normalise them ourselves. We conform to what seems to be the urn:issn
way of displaying issns. A dash after the first four digits, always a
capital X if the check digit is an X.